refactor: make lsp client available outside of this package (#77)
