+++ pkg/flink/lsp/lsp_completer.go
@@ -1,4 +1,4 @@
-package autocomplete
+package lsp
 
 import (
 	"context"
