+++ pkg/flink/lsp/lsp_completer_ws.go
@@ -1,4 +1,4 @@
-package autocomplete
+package lsp
 
 import (
 	"context"
